relocating functions view
