+++ x-pack/plugins/profiling/public/components/app/functions_view/index.tsx
@@ -8,18 +8,18 @@ import { EuiFlexGroup, EuiFlexItem, EuiPageHeaderContentProps } from '@elastic/e
 import { i18n } from '@kbn/i18n';
 import { TypeOf } from '@kbn/typed-react-router-config';
 import React from 'react';
-import { useProfilingParams } from '../../hooks/use_profiling_params';
-import { useProfilingRouter } from '../../hooks/use_profiling_router';
-import { useProfilingRoutePath } from '../../hooks/use_profiling_route_path';
-import { useTimeRange } from '../../hooks/use_time_range';
-import { useTimeRangeAsync } from '../../hooks/use_time_range_async';
-import { ProfilingRoutes } from '../../routing';
-import { AsyncComponent } from '../async_component';
-import { useProfilingDependencies } from '../contexts/profiling_dependencies/use_profiling_dependencies';
-import { PrimaryAndComparisonSearchBar } from '../primary_and_comparison_search_bar';
-import { ProfilingAppPageTemplate } from '../profiling_app_page_template';
-import { RedirectTo } from '../redirect_to';
-import { TopNFunctionsTable } from '../topn_functions';
+import { useProfilingParams } from '../../../hooks/use_profiling_params';
+import { useProfilingRouter } from '../../../hooks/use_profiling_router';
+import { useProfilingRoutePath } from '../../../hooks/use_profiling_route_path';
+import { useTimeRange } from '../../../hooks/use_time_range';
+import { useTimeRangeAsync } from '../../../hooks/use_time_range_async';
+import { ProfilingRoutes } from '../../../routing';
+import { AsyncComponent } from '../../async_component';
+import { useProfilingDependencies } from '../../contexts/profiling_dependencies/use_profiling_dependencies';
+import { PrimaryAndComparisonSearchBar } from '../../primary_and_comparison_search_bar';
+import { ProfilingAppPageTemplate } from '../../profiling_app_page_template';
+import { RedirectTo } from '../../redirect_to';
+import { TopNFunctionsTable } from '../../topn_functions';
 
 export function FunctionsView({ children }: { children: React.ReactElement }) {
   const {
