+++ x-pack/plugins/profiling/public/routing/index.tsx
@@ -13,7 +13,7 @@ import { TopNFunctionSortField, topNFunctionSortFieldRt } from '../../common/fun
 import { StackTracesDisplayOption, TopNType } from '../../common/stack_traces';
 import { FlameGraphComparisonMode, FlameGraphNormalizationMode } from '../../common/flamegraph';
 import { FlameGraphsView } from '../components/app/flame_graphs_view';
-import { FunctionsView } from '../components/functions_view';
+import { FunctionsView } from '../components/app/functions_view';
 import { RedirectTo } from '../components/redirect_to';
 import { RouteBreadcrumb } from '../components/route_breadcrumb';
 import { StackTracesView } from '../components/stack_traces_view';
